fix: small code position change
