+++ src/components/LoaderSpinner.vue
@@ -5,75 +5,22 @@
 </template>
 
 <script lang="ts">
-import { defineComponent, ref, computed } from 'vue';
+import { defineComponent } from 'vue';
 import ProgressSpinner from 'primevue/progressspinner';
-import { HTTP, HTTP_AUTH } from '@/core/api/http-common';
-import { AxiosError, AxiosRequestConfig, AxiosResponse } from 'axios';
 
 export default defineComponent({
   name: 'LoaderSpinner',
   components: {
     ProgressSpinner,
   },
+  props: {
+    isLoading: {
+      type: Boolean,
+      required: true,
+    },
+  },
   setup() {
-    const isLoading = ref(false);
-
-    computed(() => {
-      return isLoading.value;
-    });
-
-    const showLoader = () => {
-      isLoading.value = true;
-    };
-    const hideLoader = () => {
-      isLoading.value = false;
-    };
-
-    HTTP.interceptors.request.use(
-      (req: AxiosRequestConfig) => {
-        showLoader();
-        return req;
-      },
-      (error: AxiosError) => {
-        hideLoader();
-        return error;
-      }
-    );
-
-    HTTP.interceptors.response.use(
-      (response: AxiosResponse) => {
-        hideLoader();
-        return response;
-      },
-      (error: AxiosError) => {
-        hideLoader();
-        return error;
-      }
-    );
-
-    HTTP_AUTH.interceptors.request.use(
-      (req: AxiosRequestConfig) => {
-        showLoader();
-        return req;
-      },
-      (error: AxiosError) => {
-        hideLoader();
-        return error;
-      }
-    );
-
-    HTTP_AUTH.interceptors.response.use(
-      (response: AxiosResponse) => {
-        hideLoader();
-        return response;
-      },
-      (error: AxiosError) => {
-        hideLoader();
-        return error;
-      }
-    );
-
-    return { showLoader, hideLoader, isLoading };
+    return {};
   },
 });
 </script>
